requirements.txt
nose>=1.3.7
paho-mqtt>=1.5.1
pip>=2.3.1
pylint>=2.6.0
